Move functions/src/helpers to backend/shared/helpers
