+++ backend/functions/src/create-market.ts
@@ -32,9 +32,9 @@ import { NUMERIC_BUCKET_COUNT } from 'common/numeric-constants'
 import { User } from 'common/user'
 import { Group, GroupLink, MAX_ID_LENGTH } from 'common/group'
 import { getPseudoProbability } from 'common/pseudo-numeric'
-import { getCloseDate } from './helpers/openai-utils'
+import { getCloseDate } from 'shared/helpers/openai-utils'
 import { marked } from 'marked'
-import { mintAndPoolCert } from './helpers/cert-txns'
+import { mintAndPoolCert } from 'shared/helpers/cert-txns'
 import { Bet } from 'common/bet'
 import { QfAddPoolTxn } from 'common/txn'
