+++ backend/functions/src/dividend-cert.ts
@@ -4,7 +4,7 @@ import { z } from 'zod'
 import { User } from 'common/user'
 import { getDividendPayouts } from 'common/calculate/cert'
 import { Uniswap2CertContract } from 'common/contract'
-import { dividendTxns } from './helpers/cert-txns'
+import { dividendTxns } from 'shared/helpers/cert-txns'
 import { CertTxn } from 'common/txn'
 import { payUsers } from 'shared/utils'
