+++ backend/functions/src/scripts/recommend.ts
@@ -1,6 +1,6 @@
 import { getMarketRecommendations, user_data } from 'common/recommendation'
 import { sortBy } from 'lodash'
-import { readJson, writeJson } from '../helpers/file'
+import { readJson, writeJson } from 'shared/helpers/file'
 import { getContract } from 'shared/utils'
 
 import { initAdmin } from './script-init'
