+++ backend/functions/src/swap-cert.ts
@@ -4,7 +4,7 @@ import { z } from 'zod'
 import { User } from 'common/user'
 import { afterSwap } from 'common/calculate/uniswap2'
 import { Cert, Uniswap2CertContract } from 'common/contract'
-import { buyFromPool } from './helpers/cert-txns'
+import { buyFromPool } from 'shared/helpers/cert-txns'
 import { CertTxn } from 'common/txn'
 import { getCertOwnershipUsers } from 'common/calculate/cert'
