+++ backend/shared/helpers/openai-utils.ts
