+++ backend/shared/helpers/user-contract-metrics.ts
