+++ backend/shared/package.json
@@ -14,7 +14,9 @@
     "firebase-admin": "11.2.0",
     "lodash": "4.17.21",
     "node-fetch": "2",
-    "pg-promise": "11.0.2"
+    "pg-promise": "11.0.2",
+    "dayjs": "1.11.4",
+    "openai": "3.0.1"
   },
   "devDependencies": {
     "@types/node-fetch": "2.6.2"
